+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   E.g. local topology name `gates.web` corresponds to the following TOML section: `[gates.web]`.
 - `Context::group()` to get a group's address.
 
+### Fixed
+- `elfo::test::proxy`: a race condition at startup.
+
 ## [0.1.8] - 2021-05-06
 ### Added
 - `Proxy::subproxy()`